+++ src/app.jsx
@@ -4,17 +4,17 @@ import Login from "./routes/auth/login";
 import ForgotPassword from "./routes/auth/forgot";
 import Signup from "./routes/auth/signup";
 import Dashboard from "./routes/dashboard";
-import Billing from "./routes/billing";
+import Billing from "./routes/billing/billing";
 import Delivery from "./routes/delivery/delivery";
-import Product from "./routes/products";
-import Vendor from "./routes/vendors";
+import Product from "./routes/products/products";
+import Vendor from "./routes/vendors/vendors";
 import Homepage from "./routes/landingpage";
 import Profile from "./routes/profile";
 import Notifications from "./routes/notifications";
 import Faq from "./routes/faqs";
-import AddVendor from "./routes/addVendor";
-import AddProduct from "./routes/addProduct";
-import AddBill from "./routes/addBill";
+import AddVendor from "./routes/vendors/addVendor";
+import AddProduct from "./routes/products/addProduct";
+import AddBill from "./routes/billing/addBill";
 import { PersistGate } from "redux-persist/integration/react";
 import { Provider } from "react-redux";
 import { persistor, store } from "./redux";
