+++ src/routes/billing/billing.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import "../index.css";
-import Header from "../components/header";
-import Sidebar from "../components/sidebar";
+import "../../index.css";
+import Header from "../../components/header";
+import Sidebar from "../../components/sidebar";
 import { PieChart } from "react-minimal-pie-chart";
 
 const invoices = [
