+++ src/routes/products/addProduct.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import "../index.css";
-import Header from "../components/header";
-import Sidebar from "../components/sidebar";
+import "../../index.css";
+import Header from "../../components/header";
+import Sidebar from "../../components/sidebar";
 import { Link } from "react-router-dom";
 
 const AddProducts = () => {
